Header done and buttons created
Button component has been created, typescript for changing appearance has been applied through input and output. Thus the label can be chosen aswell, event Emitter has been added for functions aswe;ll. Card component for services was created and data is going to be made in a model soon and setup in ts file, this will then mean that instead of having 3 compnents in html, I can loop through it and create it. Furthemore, animations have been added in the header and button added, something feels off but I will look into that throughout the term, some bugs with the slider but a temporary solution was found